Fix up ProjectConfigurationTab
