+++ server/src/main/java/com/octopus/teamcity/opentelemetry/server/endpoints/IOTELEndpointHandler.java
@@ -1,16 +1,11 @@
 package com.octopus.teamcity.opentelemetry.server.endpoints;
 
-import com.octopus.teamcity.opentelemetry.server.HeaderDto;
-import com.octopus.teamcity.opentelemetry.server.ProjectConfigurationSettingsController;
 import com.octopus.teamcity.opentelemetry.server.SetProjectConfigurationSettingsRequest;
 import io.opentelemetry.sdk.trace.SpanProcessor;
-import jetbrains.buildServer.controllers.ActionErrors;
 import jetbrains.buildServer.serverSide.SBuild;
 import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.Map;
 
 public interface IOTELEndpointHandler {
@@ -19,4 +14,6 @@ public interface IOTELEndpointHandler {
     SpanProcessor BuildSpanProcessor(String endpoint, Map<String, String> params);
 
     SetProjectConfigurationSettingsRequest GetSetProjectConfigurationSettingsRequest(HttpServletRequest request);
+
+    void mapParamsToModel(Map<String, String> params, Map<String, Object> model);
 }
