+++ server/src/main/java/com/octopus/teamcity/opentelemetry/server/endpoints/custom/CustomOTELEndpointHandler.java
@@ -7,18 +7,20 @@ import io.opentelemetry.exporter.otlp.trace.OtlpGrpcSpanExporterBuilder;
 import io.opentelemetry.sdk.trace.SpanProcessor;
 import io.opentelemetry.sdk.trace.export.BatchSpanProcessor;
 import io.opentelemetry.sdk.trace.export.SpanExporter;
-import jetbrains.buildServer.controllers.ActionErrors;
 import jetbrains.buildServer.serverSide.SBuild;
 import jetbrains.buildServer.serverSide.crypt.EncryptUtil;
+import jetbrains.buildServer.serverSide.crypt.RSACipher;
 import jetbrains.buildServer.web.openapi.PluginDescriptor;
 import org.apache.log4j.Logger;
 import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
 
-import static com.octopus.teamcity.opentelemetry.common.PluginConstants.PROPERTY_KEY_HEADERS;
+import static com.octopus.teamcity.opentelemetry.common.PluginConstants.*;
+import static com.octopus.teamcity.opentelemetry.common.PluginConstants.PROPERTY_KEY_HONEYCOMB_APIKEY;
 
 public class CustomOTELEndpointHandler implements IOTELEndpointHandler {
     private final PluginDescriptor pluginDescriptor;
@@ -50,6 +52,25 @@ public class CustomOTELEndpointHandler implements IOTELEndpointHandler {
     @Override
     public SetProjectConfigurationSettingsRequest GetSetProjectConfigurationSettingsRequest(HttpServletRequest request) {
         return new SetCustomProjectConfigurationSettingsRequest(request);
+    }
+
+    @Override
+    public void mapParamsToModel(Map<String, String> params, Map<String, Object> model) {
+        model.put("otelEndpoint", params.get(PROPERTY_KEY_ENDPOINT));
+
+        var headers = new ArrayList<HeaderDto>();
+        params.forEach((k,v)->{
+            if (k.startsWith(PROPERTY_KEY_HEADERS)) {
+                var key = k.substring(PROPERTY_KEY_HEADERS.length());
+                key = key.substring(1, key.length() - 1);
+                var header = new HeaderDto(key,
+                        EncryptUtil.isScrambled(v) ? EncryptUtil.unscramble(v) : v,
+                        EncryptUtil.isScrambled(v) ? "password" : "plaintext");
+                headers.add(header);
+            }
+        });
+
+        model.put("otelHeaders", headers);
     }
 
     private SpanProcessor buildGrpcSpanProcessor(Map<String, String> headers, String exporterEndpoint) {
